% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_scan.R
\name{gtfs_scan}
\alias{gtfs_scan}
\title{Scan the position of all vehicles}
\usage{
gtfs_scan(dir, day, time)
}
\arguments{
\item{dir}{GTFS directory (in .zip format)}

\item{day}{Day ("YYYY-MM-DD")}

\item{time}{Time ("HH:MM:SS")}
}
\value{
Generates a leaflet map representing the position of vehicles and the 'position_vehicles' object in 'sf' format.
}
\description{
Determines the geographical position of vehicles according to a timetable, based on GTFS data.
}
\examples{
\dontrun{
gtfs_scan(
  dir = "C/.../GTFS.zip",
  day = "2025-11-05",
  time = "08:05:00")
}
}
